Fixed float32 parameter bug.

+++ leap_c/ocp/acados/utils/prepare_solver.py
@@ -61,16 +61,16 @@ def prepare_batch_solver(
     ):
         # if p_stagewise is None and default exist, load default p
         param_default = ocp.model.p.tile(batch_size, N + 1)  # type:ignore
-        param = param_default.reshape(batch_size, -1)
+        param = param_default.reshape(batch_size, -1).astype(np.float64)
         batch_solver.set_flat("p", param)
     elif p_stagewise is not None and p_stagewise_sparse_idx is None:
         # if p_stagewise is provided, set it
-        param = p_stagewise.reshape(batch_size, -1)
+        param = p_stagewise.reshape(batch_size, -1).astype(np.float64)
         batch_solver.set_flat("p", param)
     elif p_stagewise is not None and p_stagewise_sparse_idx is not None:
         # if p_stagewise is provided and sparse indices are provided, set it
         for idx, stage in product(range(batch_size), range(N + 1)):
-            param = p_stagewise[idx, stage, :]
+            param = p_stagewise[idx, stage, :].astype(np.float64)
             solver = batch_solver.ocp_solvers[idx]
             solver.set_params_sparse(stage, param, p_stagewise_sparse_idx)
